+++ Reflection/PropertyClass.h
@@ -46,6 +46,8 @@
 #include "Reflection/TypeCopyBits.h"
 #include "MedusaDll.h"
 
+#include <cstdint>
+
 //---------------------------------------------------------------------------------------------------
 
 template<typename T>
@@ -54,7 +56,7 @@ void AddProperty( PropertyList * pList, const char * pName, const T & member,
 {
 	Type * pType = TypeSerialized<T>::instance();
 	pList->addProperty( new Property( new StaticContainer( nElements, pType->size() ), pName, 
-		reinterpret_cast<int>( &member ), pType, nFlags, pNote ) );
+		(int)reinterpret_cast<uintptr_t>( &member ), pType, nFlags, pNote ) );
 }
 
 template<typename T>
@@ -64,7 +66,7 @@ void AddProperty( PropertyList * pList, const char * pName, const Array<T> & mem
 	Type * pType = TypeSerialized< T >::instance();
 	TypeSerialized< Array<T> >::instance();
 	pList->addProperty( new Property( new ArrayContainer<T>(), pName, 
-		reinterpret_cast<int>( &member ), pType, nFlags, pNote ) );
+		(int)reinterpret_cast<uintptr_t>( &member ), pType, nFlags, pNote ) );
 }
 
 template<typename T>
@@ -74,7 +76,7 @@ void AddProperty( PropertyList * pList, const char * pName, const std::vector<T>
 	Type * pType = TypeSerialized< T >::instance();
 	TypeSerialized< Array<T> >::instance();
 	pList->addProperty( new Property( new VectorContainer<T>(), pName, 
-		reinterpret_cast<int>( &member ), pType, nFlags, pNote ) );
+		(int)reinterpret_cast<uintptr_t>( &member ), pType, nFlags, pNote ) );
 }
 
 template<typename T>
@@ -84,7 +86,7 @@ void AddProperty( PropertyList * pList, const char * pName, const std::list<T> &
 	Type * pType = TypeSerialized< T >::instance();
 	TypeSerialized< Array<T> >::instance();
 	pList->addProperty( new Property( new ListContainer<T>(), pName, 
-		reinterpret_cast<int>( &member ), pType, nFlags, pNote ) );
+		(int)reinterpret_cast<uintptr_t>( &member ), pType, nFlags, pNote ) );
 }
 
 template<typename T>
@@ -95,7 +97,7 @@ void AddEnumProperty( PropertyList * pList, const char * pName, const T & member
 	Type * pUnsignedInt = TypeSerialized<unsigned int>::instance();
 
 	Property * pProperty = new Property( new StaticContainer( nElements, pType->size() ), pName, 
-		reinterpret_cast<int>( &member ), pType, nFlags, pNote );
+		(int)reinterpret_cast<uintptr_t>( &member ), pType, nFlags, pNote );
 	pList->addProperty( pProperty );
 
 	// T -> T
@@ -118,7 +120,7 @@ void AddBitsProperty( PropertyList * pList, const char * pName, const T & member
 	Type * pDWORD = TypeSerialized<dword>::instance();
 
 	Property * pProperty = new Property( new StaticContainer( nElements, pType->size() ), pName, 
-		reinterpret_cast<int>( &member ), pType, nFlags, pNote );
+		(int)reinterpret_cast<uintptr_t>( &member ), pType, nFlags, pNote );
 	pList->addProperty( pProperty );
 
 	// T -> T
@@ -168,8 +170,8 @@ public:
 		static PropertyListTemplate<CLASS> LIST( BASE::staticPropertyList() );	\
 		if (! LIST.initialized() )										\
 		{																\
-			int nOffset = ((int)(CLASS *)(PropertyClass *)1) - 1;		\
-			CLASS * pOBJECT = (CLASS *)nOffset;							\
+			int nOffset = (int)(uintptr_t)(CLASS *)(PropertyClass *)1 - 1;	\
+			CLASS * pOBJECT = (CLASS *)(uintptr_t)nOffset;				\
 			LIST.initialize();
 
 
@@ -189,8 +191,8 @@ public:
 		static PropertyListAbstractTemplate<CLASS> LIST( BASE::staticPropertyList() );	\
 		if (! LIST.initialized() )										\
 		{																\
-			int nOffset = ((int)(CLASS *)(PropertyClass *)1) - 1;		\
-			CLASS * pOBJECT = (CLASS *)nOffset;							\
+			int nOffset = (int)(uintptr_t)(CLASS *)(PropertyClass *)1 - 1;	\
+			CLASS * pOBJECT = (CLASS *)(uintptr_t)nOffset;				\
 			LIST.initialize();
 
 #define BEGIN_ABSTRACT_PROPERTY_LIST( CLASS, BASE )						\
